requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
sqlmodel==0.0.8
pytest==7.4.0
httpx==0.24.0
chromadb==0.4.10
sentence-transformers==2.2.2
python-dotenv==1.0.0
openai==0.27.8
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3
